requirements.txt
requests==2.23.0
botbuilder-integration-aiohttp>=4.13.0
pykube-ng==21.3.0